+++ api/admin-api/src/main/java/cc/uncarbon/module/adminapi/AdminApiApplication.java
@@ -1,0 +1,13 @@
+package cc.uncarbon.module.adminapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication
+public class AdminApiApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(AdminApiApplication.class, args);
+    }
+
+}
