+++ api/admin-api/src/main/java/cc/uncarbon/module/adminapi/config/CorsConfiguration.java
@@ -1,0 +1,32 @@
+package cc.uncarbon.module.adminapi.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+/**
+ * Spring Boot 2.4+ 跨域设置需要变化
+ * 参考https://blog.csdn.net/jxysgzs/article/details/110818712
+ *
+ * 因可能个别项目需要特别的跨域设置，单拎出来放在这
+ *
+ * @author Uncarbon
+ */
+@Configuration
+public class CorsConfiguration implements WebMvcConfigurer {
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry
+                // 设置允许跨域的路由规则
+                .addMapping("/**")
+                // 设置允许跨域请求的域名
+                .allowedOriginPatterns("*")
+                // 是否允许证书（cookies）
+                .allowCredentials(true)
+                // 设置允许的方法
+                .allowedMethods("*")
+                // 跨域允许时间
+                .maxAge(3600);
+    }
+}
