--- api/saas-api/src/main/java/cc/uncarbon/module/saasapi/SaasApiApplication.java
@@ -1,13 +1,0 @@
-package cc.uncarbon.module.saasapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-@SpringBootApplication
-public class SaasApiApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(SaasApiApplication.class, args);
-    }
-
-}
