--- api/saas-api/src/main/java/cc/uncarbon/module/saasapi/web/sys/AdminSysLogController.java
@@ -1,50 +1,0 @@
-package cc.uncarbon.module.saasapi.web.sys;
-
-import cc.uncarbon.framework.core.constant.HelioConstant;
-import cc.uncarbon.framework.core.page.PageParam;
-import cc.uncarbon.framework.core.page.PageResult;
-import cc.uncarbon.framework.web.model.response.ApiResult;
-import cc.uncarbon.module.sys.facade.SysLogFacade;
-import cc.uncarbon.module.sys.model.request.AdminListSysLogDTO;
-import cc.uncarbon.module.sys.model.response.SysLogBO;
-import cc.uncarbon.module.saasapi.util.AdminStpUtil;
-import cn.dev33.satoken.annotation.SaCheckLogin;
-import cn.dev33.satoken.annotation.SaCheckPermission;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.dubbo.config.annotation.DubboReference;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-
-@SaCheckLogin(type = AdminStpUtil.TYPE)
-@Slf4j
-@Api(value = "系统日志管理接口", tags = {"系统日志管理接口"})
-@RequestMapping("/api/v1")
-@RestController
-public class AdminSysLogController {
-
-    private static final String PERMISSION_PREFIX = "SysLog:";
-
-    @DubboReference(version = HelioConstant.Version.DUBBO_VERSION_V1, validation = HelioConstant.Dubbo.ENABLE_VALIDATION)
-    private SysLogFacade sysLogFacade;
-
-
-    @SaCheckPermission(type = AdminStpUtil.TYPE, value = PERMISSION_PREFIX + HelioConstant.Permission.RETRIEVE)
-    @ApiOperation(value = "分页列表")
-    @GetMapping(value = "/sys/logs")
-    public ApiResult<PageResult<SysLogBO>> list(PageParam pageParam, AdminListSysLogDTO dto) {
-        return ApiResult.data(sysLogFacade.adminList(pageParam, dto));
-    }
-
-    @SaCheckPermission(type = AdminStpUtil.TYPE, value = PERMISSION_PREFIX + HelioConstant.Permission.RETRIEVE)
-    @ApiOperation(value = "详情")
-    @GetMapping(value = "/sys/logs/{id}")
-    public ApiResult<SysLogBO> getById(@PathVariable Long id) {
-        return ApiResult.data(sysLogFacade.getOneById(id, true));
-    }
-
-}
