--- knife4j/aggregation/src/main/java/cc/uncarbon/module/aggregation/Knife4jAggregationApplication.java
@@ -1,13 +1,0 @@
-package cc.uncarbon.module.aggregation;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-@SpringBootApplication
-public class Knife4jAggregationApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(Knife4jAggregationApplication.class, args);
-    }
-
-}
